Refactor wallet connect classes
- Delete unused files
- Change allowed blockchains types
- Rename classes
