+++ UnstoppableWallet/UnstoppableWallet/Modules/WalletConnect/Request/Sign/WalletConnectSignMessageRequestModule.swift
@@ -3,7 +3,7 @@ import EthereumKit
 
 struct WalletConnectSignMessageRequestModule {
 
-    static func viewController(signService: WalletConnectV1XMainService, requestId: Int) -> UIViewController? {
+    static func viewController(signService: WalletConnectV1MainService, requestId: Int) -> UIViewController? {
         guard let request = signService.pendingRequest(requestId: requestId) as? WalletConnectSignMessageRequest else {
             return nil
         }
