+++ UnstoppableWallet/UnstoppableWallet/Modules/WalletConnect/ScanQr/WalletConnectScanQrViewModel.swift
@@ -1,37 +1,18 @@
 import RxSwift
-import RxRelay
 import RxCocoa
-import WalletConnectV1
 
 class WalletConnectScanQrViewModel {
-    private let service: WalletConnectService
-
-    private let openMainRelay = PublishRelay<Void>()
     private let openErrorRelay = PublishRelay<Error>()
 
-    init(service: WalletConnectService) {
-        self.service = service
+    init() {
     }
 
 }
 
 extension WalletConnectScanQrViewModel {
 
-    var openMainSignal: Signal<Void> {
-        openMainRelay.asSignal()
-    }
+    func didScan(string: String) {
 
-    var openErrorSignal: Signal<Error> {
-        openErrorRelay.asSignal()
-    }
-
-    func handleScanned(string: String) {
-        do {
-            try service.connect(uri: string)
-            openMainRelay.accept(())
-        } catch {
-            openErrorRelay.accept(error)
-        }
     }
 
 }
