+++ UnstoppableWallet/UnstoppableWallet/Modules/WalletConnect/WalletConnectSessionKiller.swift
