--- UnstoppableWallet/UnstoppableWallet/Modules/WalletConnectX/ScanQr/WalletConnectXScanQrViewController.swift
@@ -1,17 +1,0 @@
-import ThemeKit
-
-class WalletConnectXScanQrViewController: ScanQrViewController {
-
-    override func onScan(string: String) {
-        delegate?.didScan(viewController: self, string: string)
-    }
-
-}
-
-extension WalletConnectXScanQrViewController: IWalletConnectErrorDelegate {
-
-    func onDismiss() {
-        startCaptureSession()
-    }
-
-}
