--- UnstoppableWallet/UnstoppableWallet/Modules/WalletConnectX/ScanQr/WalletConnectXScanQrViewModel.swift
@@ -1,18 +1,0 @@
-import RxSwift
-import RxCocoa
-
-class WalletConnectXScanQrViewModel {
-    private let openErrorRelay = PublishRelay<Error>()
-
-    init() {
-    }
-
-}
-
-extension WalletConnectXScanQrViewModel {
-
-    func didScan(string: String) {
-
-    }
-
-}
