--- UnstoppableWallet/UnstoppableWallet/Modules/WalletConnectX/WalletConnectXModule.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol IWalletConnectSignService {
-    func approveRequest(id: Int, result: Data)
-    func rejectRequest(id: Int)
-}
